TST: reorganize tools.tests
